requirements.txt
langchain==0.0.123
openai==0.27.2
qdrant-client
numpy
pandas
gdown
python-dotenv